pyproject.toml
[build-system]
requires = [
  "setuptools >= 77.0.3",
  "Cython >= 3.1",
  "numpy >= 1.20",
]
build-backend = "setuptools.build_meta"

[project]
name = "gerbls"
dynamic = ["version"]
authors = [{name = "Kristo Ment", email = "kristo.ment@gmail.com"}]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">= 3.9"

classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: C",
  "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
  "numpy >= 1.20",
  "scipy",
]

[project.optional-dependencies]
extras = [
  "batman-package",
  "matplotlib", 
  "pytest",
]

[project.urls]
Documentation = "https://gerbls.readthedocs.io/"
Repository = "https://github.com/kment/GERBLS.git"

[tool.setuptools.dynamic]
version = {attr = "gerbls.__version__"}

[tool.setuptools.packages.find]
where = ["src"]